Add option to CMake to build as static library
