+++ CMakeLists.txt
@@ -19,6 +19,7 @@ ENDIF()
 OPTION(USING_MALLOC_TRIM "Call malloc_trim after processing request to lower memory usage (Your system must support malloc_trim)." OFF)
 #now using internal MD5 calculation
 #OPTION(USING_MBEDTLS "Use mbedTLS instead of OpenSSL for MD5 calculation." OFF)
+OPTION(BUILD_STATIC_LIBRARY "Build a static library containing only the essential part." OFF)
 
 INCLUDE(CheckCXXSourceCompiles)
 CHECK_CXX_SOURCE_COMPILES(
@@ -28,14 +29,18 @@ void main(){std::to_string(0);}
 " HAVE_TO_STRING)
 
 IF(HAVE_TO_STRING)
-    ADD_DEFINITIONS(-DHAVE_TO_STRING)
+    TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} -DHAVE_TO_STRING)
 ENDIF()
 
 IF(USING_MALLOC_TRIM)
-    ADD_DEFINITIONS(-DMALLOC_TRIM)
+    TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} -DMALLOC_TRIM)
 ENDIF()
 
-ADD_EXECUTABLE(subconverter 
+SET(BUILD_TARGET_NAME ${PROJECT_NAME})
+
+IF(NOT BUILD_STATIC_LIBRARY)
+
+ADD_EXECUTABLE(${BUILD_TARGET_NAME} 
     src/generator/config/nodemanip.cpp
     src/generator/config/ruleconvert.cpp
     src/generator/config/subexport.cpp
@@ -61,62 +66,101 @@ ADD_EXECUTABLE(subconverter
     src/utils/string.cpp
     src/utils/system.cpp
     src/utils/urlencode.cpp)
-INCLUDE_DIRECTORIES(src)
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} src)
 LINK_DIRECTORIES(${CMAKE_SOURCE_DIR})
 
 FIND_PACKAGE(PkgConfig REQUIRED)
 
 SET(THREADS_PREFER_PTHREAD_FLAG ON)
 FIND_PACKAGE(Threads REQUIRED)
-TARGET_LINK_LIBRARIES(subconverter ${CMAKE_THREAD_LIBS_INIT})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME}  ${CMAKE_THREAD_LIBS_INIT})
 
 PKG_CHECK_MODULES(LIBEVENT libevent>=2.1.10 REQUIRED)
 FIND_PATH(LIBEVENT_INCLUDE_DIR NAMES event.h PATHS ${LIBEVENT_INCLUDE_DIRS})
 FIND_LIBRARY(LIBEVENT_LIBRARY NAMES event PATHS ${LIBEVENT_LIBRARY_DIRS})
-LINK_DIRECTORIES(${LIBEVENT_LIBRARY_DIRS})
-INCLUDE_DIRECTORIES(${LIBEVENT_INCLUDE_DIR})
-TARGET_LINK_LIBRARIES(subconverter ${LIBEVENT_LIBRARY})
+TARGET_LINK_DIRECTORIES(${BUILD_TARGET_NAME} ${LIBEVENT_LIBRARY_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${LIBEVENT_INCLUDE_DIR})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} ${LIBEVENT_LIBRARY})
 
 FIND_PACKAGE(CURL 7.54.0 REQUIRED)
-LINK_DIRECTORIES(${CURL_LIBRARY_DIRS})
-INCLUDE_DIRECTORIES(${CURL_INCLUDE_DIRS})
-TARGET_LINK_LIBRARIES(subconverter CURL::libcurl)
-ADD_DEFINITIONS(-DCURL_STATICLIB)
+TARGET_LINK_DIRECTORIES(${BUILD_TARGET_NAME} ${CURL_LIBRARY_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${CURL_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} CURL::libcurl)
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} -DCURL_STATICLIB)
 
 FIND_PACKAGE(Rapidjson REQUIRED)
-INCLUDE_DIRECTORIES(${RAPIDJSON_INCLUDE_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${RAPIDJSON_INCLUDE_DIRS})
 
 FIND_PACKAGE(toml11 REQUIRED)
-INCLUDE_DIRECTORIES(${TOML11_INCLUDE_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${TOML11_INCLUDE_DIRS})
 
 PKG_CHECK_MODULES(YAML_CPP yaml-cpp>=0.6.3 REQUIRED)
 FIND_LIBRARY(YAML_CPP_LIBRARY NAMES yaml-cpp yaml-cppd PATHS ${YAML_CPP_LIBRARY_DIRS})
-LINK_DIRECTORIES(${YAML_CPP_LIBRARY_DIRS})
-INCLUDE_DIRECTORIES(${YAML_CPP_INCLUDE_DIRS})
-TARGET_LINK_LIBRARIES(subconverter ${YAML_CPP_LIBRARY})
-ADD_DEFINITIONS(-DYAML_CPP_STATIC_DEFINE)
+TARGET_LINK_DIRECTORIES(${BUILD_TARGET_NAME} ${YAML_CPP_LIBRARY_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${YAML_CPP_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} ${YAML_CPP_LIBRARY})
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} -DYAML_CPP_STATIC_DEFINE)
 
-#IF(USING_STD_REGEX STREQUAL "ON")
-#    ADD_DEFINITIONS(-DUSE_STD_REGEX)
-#ELSE()
-    FIND_PACKAGE(PCRE2 REQUIRED)
-    INCLUDE_DIRECTORIES(${PCRE2_INCLUDE_DIRS})
-    TARGET_LINK_LIBRARIES(subconverter ${PCRE2_LIBRARY})
-    ADD_DEFINITIONS(-DPCRE2_STATIC)
-#ENDIF()
+FIND_PACKAGE(PCRE2 REQUIRED)
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${PCRE2_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} ${PCRE2_LIBRARY})
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} -DPCRE2_STATIC)
 
 FIND_PACKAGE(QuickJS REQUIRED)
-INCLUDE_DIRECTORIES(${QUICKJS_INCLUDE_DIRS})
-TARGET_LINK_LIBRARIES(subconverter ${QUICKJS_LIBRARIES})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${QUICKJS_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} ${QUICKJS_LIBRARIES})
 
 FIND_PACKAGE(LibCron REQUIRED)
-INCLUDE_DIRECTORIES(${LIBCRON_INCLUDE_DIRS})
-TARGET_LINK_LIBRARIES(subconverter ${LIBCRON_LIBRARIES})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} ${LIBCRON_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} ${LIBCRON_LIBRARIES})
 
 IF(WIN32)
-    TARGET_LINK_LIBRARIES(subconverter wsock32 ws2_32)
+    TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} wsock32 ws2_32)
 ELSE()
     INCLUDE(GNUInstallDirs)
-    INSTALL(TARGETS subconverter DESTINATION ${CMAKE_INSTALL_BINDIR}/subconverter)
-    INSTALL(DIRECTORY base/ DESTINATION ${CMAKE_INSTALL_BINDIR}/subconverter FILES_MATCHING PATTERN "*")
+    INSTALL(TARGETS ${BUILD_TARGET_NAME} DESTINATION ${CMAKE_INSTALL_BINDIR}/${BUILD_TARGET_NAME})
+    INSTALL(DIRECTORY base/ DESTINATION ${CMAKE_INSTALL_BINDIR}/${BUILD_TARGET_NAME} FILES_MATCHING PATTERN "*")
+ENDIF()
+
+ELSE()
+
+ADD_LIBRARY(${BUILD_TARGET_NAME} STATIC
+    src/generator/config/ruleconvert.cpp
+    src/generator/config/subexport.cpp
+    src/generator/template/templates.cpp
+    src/lib/wrapper.cpp
+    src/parser/subparser.cpp
+    src/utils/base64/base64.cpp
+    src/utils/codepage.cpp
+    src/utils/logger.cpp
+    src/utils/md5/md5.cpp
+    src/utils/network.cpp
+    src/utils/regexp.cpp
+    src/utils/string.cpp
+    src/utils/urlencode.cpp)
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} PRIVATE -DNO_JS_RUNTIME)
+
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} PUBLIC src)
+
+FIND_PACKAGE(Rapidjson REQUIRED)
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} PRIVATE ${RAPIDJSON_INCLUDE_DIRS})
+
+FIND_PACKAGE(PkgConfig REQUIRED)
+
+PKG_CHECK_MODULES(YAML_CPP yaml-cpp>=0.6.3 REQUIRED)
+FIND_LIBRARY(YAML_CPP_LIBRARY NAMES yaml-cpp yaml-cppd PATHS ${YAML_CPP_LIBRARY_DIRS})
+TARGET_LINK_DIRECTORIES(${BUILD_TARGET_NAME} PRIVATE ${YAML_CPP_LIBRARY_DIRS})
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} PRIVATE ${YAML_CPP_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} PRIVATE ${YAML_CPP_LIBRARY})
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} PRIVATE -DYAML_CPP_STATIC_DEFINE)
+
+FIND_PACKAGE(PCRE2 REQUIRED)
+TARGET_INCLUDE_DIRECTORIES(${BUILD_TARGET_NAME} PRIVATE ${PCRE2_INCLUDE_DIRS})
+TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} PRIVATE ${PCRE2_LIBRARY})
+TARGET_COMPILE_DEFINITIONS(${BUILD_TARGET_NAME} PRIVATE -DPCRE2_STATIC)
+
+IF(WIN32)
+    TARGET_LINK_LIBRARIES(${BUILD_TARGET_NAME} PRIVATE ws2_32)
+ENDIF()
+
 ENDIF()
